requirements.txt
imageio==2.6.1
paddlepaddle
Shapely==1.7.0
Flask==1.1.1
Flask_request_params==0.3.0
six==1.14.0
pyclipper==1.2.0
matplotlib==3.1.3
scipy==1.4.1
imgaug==0.4.0
opencv-python==4.3.0.36
lmdb==1.0.0
PyMySQL==0.10.1
torch==1.4.0+cu92
torchvision==0.5.0
Flask_Cors==3.0.9
numpy==1.16.6
requests==2.24.0
paddle==1.0.2
Pillow==8.0.1
plumbum==1.6.9
